Changed units for settings. Make everything prefer in seconds rather than in samples

+++ main/cli_utils.cpp
@@ -19,10 +19,10 @@ void parse_arguments(Settings &settings, int argc, char **argv, bool &exit, bool
     arguments.push_back(new SettingArg<int>(basic_args, "tempo_window", TEMPO_WINDOW_DESC, {"tempo_window"}, settings.tempo_window));
     arguments.push_back(new SettingArg<int>(basic_args, "ref_tempo", REF_TEMPO_DESC, {"ref_tempo"}, settings.ref_tempo));
     arguments.push_back(new SettingArg<int>(basic_args, "octave_divider", OCTAVE_DIVIDER_DESC, {"octave_divider"}, settings.octave_divider));
-    arguments.push_back(new SettingArg<int>(basic_args, "smooth_length", SMOOTH_LENGTH_DESC, {"smooth_length"}, settings.smooth_length));
+    arguments.push_back(new SettingArg<double>(basic_args, "smooth_length", SMOOTH_LENGTH_DESC, {"smooth_length"}, settings.smooth_length));
     arguments.push_back(new SettingArg<float>(basic_args, "triplet_weight", TRIPLET_WEIGH_DESC, {"triplet_weight"}, settings.triplet_weight));
-    arguments.push_back(new SettingArg<int>(basic_args, "min_section_length", MIN_SECTION_LENGTH_DESC, {"min_section_length"}, settings.min_section_length));
-    arguments.push_back(new SettingArg<int>(basic_args, "max_section_length", MAX_SECTION_LENGTH_DESC, {"max_section_length"}, settings.max_section_length));
+    arguments.push_back(new SettingArg<double>(basic_args, "min_section_length", MIN_SECTION_LENGTH_DESC, {"min_section_length"}, settings.min_section_length));
+    arguments.push_back(new SettingArg<double>(basic_args, "max_section_length", MAX_SECTION_LENGTH_DESC, {"max_section_length"}, settings.max_section_length));
     arguments.push_back(new SettingArg<float>(basic_args, "bpm_doubt_window", BPM_DOUBT_WINDOW_DESC, {"bpm_doubt_window"}, settings.bpm_doubt_window));
     arguments.push_back(new SettingArg<float>(basic_args, "bpm_doubt_step", BPM_DOUBT_STEP_DESC, {"bpm_doubt_step"}, settings.bpm_doubt_step));
     arguments.push_back(new SettingArg<bool>(basic_args, "generate_click_track", GENERATE_CLICK_TRACK_DESC, {"generate_click_track", 'c'}, settings.generate_click_track));
